* Integrate new route dom * Add about us page
