+++ package.json
@@ -14,7 +14,7 @@
     "react-dom": "^18.2.0",
     "react-icons": "^4.4.0",
     "react-redux": "7.2.3",
-    "react-router-dom": "^5.3.4",
+    "react-router-dom": "6.0.2",
     "react-scripts": "5.0.1",
     "redux": "^4.0.5",
     "regenerator-runtime": "^0.13.11",
