+++ src/App.js
@@ -1,10 +1,20 @@
 import useFetchCategory from 'Hooks/useFetchCategory';
+import AboutUs from 'pages/AboutUs';
+import Home from 'pages/Home';
 import React from 'react';
-import Home from './pages/Home';
+import { Routes, Route } from 'react-router-dom';
+import Layout from './pages/Layout';
 
 function App() {
   useFetchCategory();
-  return <Home />;
+  return (
+    <Layout>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/about-us" element={<AboutUs />} />
+      </Routes>
+    </Layout>
+  );
 }
 
 export default App;
