+++ src/pages/AboutUs.js
@@ -1,13 +1,10 @@
 import React from 'react';
 
-// eslint-disable-next-line import/no-cycle
-import Layout from './Layout';
-
 function AboutUs() {
   return (
-    <Layout>
+    <div>
       <p data-testid="about-us">About Us</p>
-    </Layout>
+    </div>
   );
 }
